+++ source/Python/omegatomo/projector/proj.py
@@ -668,7 +668,10 @@ class projectorClass:
                 else:
                     self.SinM = np.flip(self.SinM, axis=2)
             
-            self.n_rays_transaxial = self.nRays
+            if 'n_rays_transaxial' not in self.__dict__:
+                self.n_rays_transaxial = int(np.sqrt(self.nRays))
+            if 'n_rays_axial' not in self.__dict__:
+                self.n_rays_axial = int(np.sqrt(self.nRays))
             self.NSinos = self.nProjections
             self.TotSinos = self.nProjections
             self.det_per_ring = self.nRowsD * self.nProjections
@@ -1213,7 +1216,7 @@ class projectorClass:
                 raise ValueError(f'DetectorVector contains an index outside the available detector heads [0, {int(self.nHeads) - 1}].')
 
             if self.projector_type in [1, 11, 12, 2, 21, 22, 3, 13, 23, 33, 31, 32]:
-                compact_ray_shift_size = 2 * int(self.nRays) * int(self.nRowsD) * int(self.nColsD) * int(self.nHeads)
+                compact_ray_shift_size = 2 * int(self.n_rays_transaxial) * int(self.n_rays_axial) * int(self.nRowsD) * int(self.nColsD) * int(self.nHeads)
                 detector_size = int(np.size(self.rayShiftsDetector))
                 source_size = int(np.size(self.rayShiftsSource))
                 if detector_size > 0 and detector_size != compact_ray_shift_size:
@@ -1231,7 +1234,7 @@ class projectorClass:
             print("Axial FOV is too small, crystal ring(s) on the boundary have no slices!")
         if self.SPECT and math.sqrt(self.nRays) % 1 != 0:
             raise ValueError("With SPECT, options.nRays has to be a square")
-        if self.SPECT and self.projector_type in [2, 12, 21, 22] and self.nRays > 1:
+        if self.SPECT and self.projector_type in [2, 12, 21, 22] and self.n_rays_transaxial * self.n_rays_axial > 1:
             print('Orthogonal distance ray tracer should be used with 1 ray.')
         if not(self.PDHG or self.PDHGKL or self.PDHGL1 or self.PDDY or self.PKMA or self.FISTA or self.FISTAL1 or self.MBSREM or self.SPS or self.MRAMLA) and any(self.precondTypeImage):
             print("Image-based preconditioning selected, but the selected algorithm(s) do not support preconditioning. No preconditioning will be performed.")