+++ source/m-files/setMissingValues.m
@@ -331,13 +331,17 @@ if ~isfield(options, 'nRays')
 end
 if ~isfield(options, 'n_rays_transaxial')
     if options.SPECT
-        options.n_rays_transaxial = options.nRays;
+        options.n_rays_transaxial = sqrt(options.nRays);
     else
         options.n_rays_transaxial = 1;
     end
 end
 if ~isfield(options, 'n_rays_axial')
-    options.n_rays_axial = 1;
+    if options.SPECT
+        options.n_rays_axial = sqrt(options.nRays);
+    else
+        options.n_rays_axial = 1;
+    end
 end
 if ~isfield(options, 'cpu_to_gpu_factor')
     options.cpu_to_gpu_factor = 1;