+++ source/opencl/general_opencl_functions.h
@@ -1563,7 +1563,7 @@ DEVICE void getDetectorCoordinatesSPECT(
     const CLGLOBAL float* d_rayShiftsDetector, // Ray shifts [mm]
     const CLGLOBAL float* d_rayShiftsSource, // Ray shifts [mm]
     const CLGLOBAL uint* d_detectorVector, // Detector head for each projection
-    int lorXY,
+    int lor,
     const FLOAT3 totalFOVmin, // FOV boundary including all multiresolution volumes
     const FLOAT3 totalFOVmax // FOV boundary including all multiresolution volumes
 ) {
@@ -1578,8 +1578,7 @@ DEVICE void getDetectorCoordinatesSPECT(
     id = i.z * NA; // Index of d_uv (detector panel normal vector)
     const uint detectorElement = i.x + i.y * d_size_x;
     const uint detectorHead = d_detectorVector[i.z];
-    const uint shiftBlock = detectorElement + detectorHead * d_size_x * d_sizey;
-    uint idShift = 2*lorXY + (2*N_RAYS) * shiftBlock; // Index of rayShiftsDetector
+    uint idShift = 2*lor + (2*N_RAYS) * detectorElement + detectorHead * d_size_x * d_sizey; // Index of rayShiftsDetector
 
 	const FLOAT apuX = d_uv[id]; // X component of detector panel normal vector
 	const FLOAT apuY = d_uv[id + 1]; // Y component of detector panel normal vector